requirements.txt
anthropic>=0.40.0
openai>=1.40.0
google-genai>=1.0.0
groq>=0.9.0
mistralai>=1.0.0
cohere>=5.0.0
boto3>=1.35.0
click>=8.1.0
rich>=13.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
jinja2>=3.1.0
